requirements.txt
# Django3.1
Django==3.1
psycopg2

# 他にインストールしたいものがあればここに記載する。
djangorestframework
django-cors-headers
networkx
matplotlib

# django[argon2]
# bcrypt
# djoser
# djangorestframework_simplejwt